refactor(shared): consolidate value-statements into single file (DEDUP-11)
4 × 52 LOC boilerplate files consolidated into cloud/packages/shared/src/value-statements.ts
(220 LOC). All exported symbol names preserved exactly — zero caller changes required.
